cleanup urlstate logic so it isn't a nightmare to maintain
still some work left to do though...
a nice side effect of this refactor is proper color scale limits
are now calculated when click on a dataset. this used to work
but broke when implementing this bookmarkable url feature. and i
just noticed right now it had broken (after it started working again)