fix memory leak
it appears we do need the featuresMap. If not, then queryRenderedFeatures keeps
almost exponentially returning more features on subsequent calls.
Thankfully, it seems no more triangles when doing actual size pixels... must be
a bug that was present in the old version of mapbox we were using and is not
present anymore in the latest.

also, the setting of various objects to null might be placebo, but I notice
a couple of megabytes difference when I do so.