make reference point toggle button show/hide reference point
if no custom reference point selected, it shows/hides the database reference point
by adding/removing the source and layer (because the default recoloring is based
on this reference point).

if a custom reference point is selected, it sets visibility of the layoutproperty
to hidden. that way, we don't see the reference point, but source is still there
so the reference point recoloring still works.