revert to postgresql 9.2
postgresql 16 broke on the fly recoloring inside getPoints in GeoJSONController.php.
Would cause returned data to not be in the proper order (of ascending point IDs),
causing the recoloring to produce corrupt/scrambled data.

Why? not sure. seems newer versions of postgresql changed the functionality
of unnest and order by. TODO: look into this