attempt at fixing getPoints in postgresql 16
Seems ROW_NUMBER() OVER() is where the issue was. See below quote from
https://www.jooq.org/doc/latest/manual/sql-building/table-expressions/with-ordinality/

"An emulation using a ROW_NUMBER() window function is possible. The ordering stability of such a derived table is at the mercy of the optimiser implementation, and may break "unexpectedly," derived table ordering isn't required to be stable in most RDBMS."